Refactor get playlist button to it's own component
