+++ src/BottomNav.jsx
@@ -1,14 +1,13 @@
 import React, { useContext } from 'react';
 import DataContext from './DataContext.jsx';
 import useMediaQuery from '@material-ui/core/useMediaQuery';
-import { BottomNavigation, Grid, Container, Box, Badge, Avatar, Fab } from '@material-ui/core';
+import { BottomNavigation, Grid, Container, Box, Fab } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
 import { Link } from 'react-router-dom';
 import theme from './theme.js';
 import SeedImage from './SeedImage.jsx';
-import queryString from 'query-string';
-import axios from 'axios';
 import SavePlaylist from './SavePlaylist.jsx';
+import GetPlaylist from './GetPlaylist.jsx';
 
 const useStyles = makeStyles((theme) => ({
   BottomNavigation: {
@@ -42,38 +41,10 @@ const useStyles = makeStyles((theme) => ({
 
 export default function BottomNav() {
   const classes = useStyles();
-  const { seedStack, setSeedStack, seed, setSeed, page, setPage, setPlaylist } = useContext(DataContext);
+  const { seedStack, setSeedStack, seed, setSeed, page, setPage } = useContext(DataContext);
   const smallScreen = useMediaQuery(theme.breakpoints.up('sm'))
   const mediumScreen = useMediaQuery(theme.breakpoints.up('md'));
   const largeScreen = useMediaQuery(theme.breakpoints.up('lg'));
-
-  function getPlaylist() {
-    setPage('/playlist');
-    let q = Object.assign({}, seed);
-    let artists = q.artist.map(artist => artist.id);
-    if (artists[0]) { q.seed_artists = artists.join(); }
-    delete q.artist;
-    let genres = q.genre.map(genre => genre.name);
-    if (genres[0]) { q.seed_genres = genres.join(); }
-    delete q.genre;
-    let tracks = q.track.map(track => track.id);
-    if (tracks[0]) { q.seed_tracks = tracks.join(); }
-    delete q.track;
-    q.limit = 50;
-
-    let stringifiedQuery = queryString.stringify(q);
-    axios({
-      method: 'get',
-      url: `/api/recommendations?${stringifiedQuery}`,
-      json: true
-    })
-      .then(results => {
-        setPlaylist(results.data)
-        if (results.data.length === 0) {
-          alert('No results. Please widen your parameters or add more seeds.')
-        }
-      });
-  }
 
   function removeSeed(item) {
     let seedGenre = seed[item.type].filter(obj => obj.name !== item.name);
@@ -88,21 +59,22 @@ export default function BottomNav() {
     <BottomNavigation className={classes.BottomNavigation}>
       {page === '/playlist' ?
         <SavePlaylist /> :
-        <Container disableGutters className={classes.container}>
-          <Grid container spacing={largeScreen ? 3 : 2}>
-            <Grid item xs={10} container alignItems='center' justify='center' spacing={mediumScreen ? 3 : smallScreen ? 2 : 1}>
-              {seedStack.map(item => (
-                <Grid item xs className={classes.seed} >
-                  <Box onClick={() => removeSeed(item)}>
-                    <SeedImage images={item.images} name={item.name} />
-                  </Box>
-                </Grid>
-              )
-              )}
-            </Grid>
-            <Grid item xs={2} container alignItems='center' justify='center'>
-              <Grid item xs={12}>
-                {page === '/' ?
+        page === '/controls' ?
+          <GetPlaylist /> :
+          <Container disableGutters className={classes.container}>
+            <Grid container spacing={largeScreen ? 3 : 2}>
+              <Grid item xs={10} container alignItems='center' justify='center' spacing={mediumScreen ? 3 : smallScreen ? 2 : 1}>
+                {seedStack.map(item => (
+                  <Grid item xs className={classes.seed} >
+                    <Box onClick={() => removeSeed(item)}>
+                      <SeedImage images={item.images} name={item.name} />
+                    </Box>
+                  </Grid>
+                )
+                )}
+              </Grid>
+              <Grid item xs={2} container alignItems='center' justify='center'>
+                <Grid item xs={12}>
                   <Fab
                     className={classes.button}
                     onClick={() => { setPage('/controls') }}
@@ -110,19 +82,11 @@ export default function BottomNav() {
                     disabled={!seedStack[0]}
                     variant='extended'
                     color='primary'
-                    fullWidth={true}>next</Fab> :
-                  <Fab
-                    color='primary'
-                    disabled={seedStack.length === 0}
-                    variant='extended'
-                    size='small'
-                    onClick={getPlaylist}
-                    component={Link}
-                    to={'/playlist'}>get playlist</Fab>}
+                    fullWidth={true}>next</Fab>
+                </Grid>
               </Grid>
             </Grid>
-          </Grid>
-        </Container>
+          </Container>
       }
     </BottomNavigation >
   )
